extracted common types to `types.ts`
